Fix a compiler warning and a comment error.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@958 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CLimitReadFile.cpp
@@ -52,7 +52,7 @@ s32 CLimitReadFile::read(void* buffer, u32 sizeToRead)
 	if (pos >= AreaEnd)
 		return 0;
 
-	if (pos + sizeToRead >= AreaEnd)
+	if (pos + (long)sizeToRead >= AreaEnd)
 		sizeToRead = AreaEnd - pos;
 
 	return File->read(buffer, sizeToRead);
